refactor: make import path from domains consistent
