+++ app/_actions/orders.ts
@@ -5,6 +5,9 @@ import { z } from "zod";
 
 import { requireSessionUser } from "@/actions/session";
 import {
+  cancelOrder,
+  createOrder,
+  deleteOrder,
   getAllOrdersByUserId,
   getFlightSeatClassById,
   getOrderConfirmation,
@@ -12,15 +15,10 @@ import {
   getOrderForAncillary,
   getOrderForPayment,
   getSavedPassengers,
-} from "@/domains/booking/booking.read.service";
-import {
-  cancelOrder,
-  createOrder,
-  deleteOrder,
   refundOrder,
   updateOrderAncillary,
-} from "@/domains/booking/orders.service";
-import { getUserBalance } from "@/domains/user/user.read.service";
+} from "@/domains/booking";
+import { getUserBalance } from "@/domains/user";
 import { auth } from "@/infra/auth";
 import type { ActionResult } from "@/types/common";
 import type {
