+++ app/_actions/quick-date-prices.ts
@@ -1,12 +1,12 @@
 "use server";
 
-import { getQuickDatePrices as getQuickDatePricesService } from "@/domains/flights/quick-date-price.service";
+import { getQuickDatePrices as getQuickDatePricesService } from "@/domains/flights";
 import {
   type GetQuickDatePricesParams,
   type QuickDatePrice,
 } from "@/types/dto";
 
-// export type { QuickDatePrice } from "@/domains/flights/quick-date-price.service";
+// export type { QuickDatePrice } from "@/domains/flights";
 
 export async function getQuickDatePrices(
   params: GetQuickDatePricesParams
