+++ app/_actions/user.ts
@@ -7,8 +7,9 @@ import {
   getUserBalance,
   getUserInfo,
   getUserSecurityStatus,
-} from "@/domains/user/user.read.service";
-import { rechargeBalance, updateUserInfo } from "@/domains/user/user.service";
+  rechargeBalance,
+  updateUserInfo,
+} from "@/domains/user";
 import { auth } from "@/infra/auth";
 import type { ActionResult } from "@/types/common";
 import type { UserInfo, UserSecurityStatus } from "@/types/dto";
