+++ content/docs/technical-design/seeding.mdx
@@ -769,7 +769,7 @@ https://ui-avatars.com/api/?name=CA&background=random&color=fff&size=128&bold=tr
 
 ```typescript
 // 在组件或 API 中使用
-import { getAirlineLogoUrl } from "@/domains/flights/utils/airline";
+import { getAirlineLogoUrl } from "@/domains/flights";
 
 const logoUrl = getAirlineLogoUrl(airline);
 ```
