+++ src/domains/auth/index.ts
@@ -1,9 +1,2 @@
-import { emailOTPClient, phoneNumberClient } from "better-auth/client/plugins";
-import { createAuthClient } from "better-auth/react";
-
-const authClient = createAuthClient({
-  baseURL: process.env.BETTER_AUTH_URL,
-  plugins: [phoneNumberClient(), emailOTPClient()],
-});
-
-export const { signIn, signUp, useSession } = authClient;
+export * from "./auth.repository";
+export * from "./auth.service";
