+++ src/domains/flights/flight.repository.test.ts
@@ -9,10 +9,13 @@ import { addDays } from "date-fns";
 import { fromZonedTime } from "date-fns-tz";
 import { describe, expect, it } from "vitest";
 
-import {
-  searchFlights,
-  searchRoundTripFlights,
-} from "@/domains/flights/flight.repository";
+import { searchFlights, searchRoundTripFlights } from "@/domains/flights";
+
+import { vi } from "vitest";
+
+vi.mock("server-only", () => {
+  return {}; // 返回空对象，不做任何检查
+});
 
 const toDateString = (date: Date) => date.toISOString().slice(0, 10);
