+++ src/domains/flights/index.ts
@@ -1,0 +1,7 @@
+export * from "./airline";
+export * from "./city.repository";
+export * from "./flight.repository";
+export * from "./flight-search-history.repository";
+export * from "./flight-search-history.service";
+export * from "./flight-search.service";
+export * from "./quick-date-price.service";
