+++ src/domains/notification/order-email.service.ts
@@ -1,4 +1,4 @@
-import { getOrderDetailById } from "@/domains/booking/orders.repository";
+import { getOrderDetailById } from "@/domains/booking";
 import { sendOrderConfirmationEmail } from "@/infra/notifications";
 import { transformOrderDetailToEmailData } from "@/lib/notification/transformers";
 import type { ServiceResult } from "@/types/result";
