+++ src/domains/passengers/index.ts
@@ -1,0 +1,3 @@
+export * from "./passenger-read.service";
+export * from "./passenger.repository";
+export * from "./passenger.service";
