+++ src/domains/payments/index.ts
@@ -1,0 +1,1 @@
+export * from "./payment.service";
