+++ src/domains/user/user.read.service.ts
@@ -2,4 +2,4 @@ export {
   getUserBalance,
   getUserInfo,
   getUserSecurityStatus,
-} from "@/domains/user/user.repository";
+} from "./user.repository";
