+++ src/services/payment-workflow.service.ts
@@ -4,7 +4,7 @@ import { nanoid } from "nanoid";
 import { db } from "@/db";
 import { user } from "@/db/schema";
 import { orders, payments } from "@/db/schema/orders";
-import { getOrderDetailById } from "@/domains/booking/orders.repository";
+import { getOrderDetailById } from "@/domains/booking";
 import { sendOrderConfirmationEmail } from "@/infra/notifications";
 import {
   getCurrencyValue,
